feat: extract appendEntries to rpc.go
